+++ src/Core.ts
@@ -1,15 +1,16 @@
 import type {ApiClientInterface} from './interfaces/ApiClientInterface';
+import {CoreInterface} from './interfaces/CoreInterface';
 
 interface Config<TValue> {
   [id: string]: TValue;
 }
 
-interface Service {
+interface ServiceBag {
   ApiClientService?: ApiClientInterface;
 }
 
-export class CoreService {
-  service: Service = {};
+export class Core implements CoreInterface {
+  service: ServiceBag = {};
   config: Config<any> = {
     IS_NODE:
       typeof process !== 'undefined' &&
@@ -18,6 +19,7 @@ export class CoreService {
       !process.versions.electron,
     REQUEST_HEADERS: {},
     SERVER_URL: 'https://api.drupal.com',
+    JSON_API_ENDPOINT: '/jsonapi',
   };
 
   public get(key: string): any {
@@ -30,12 +32,12 @@ export class CoreService {
   public set(
     key: string,
     value: boolean | number | string | {[key: string]: unknown;},
-  ): CoreService {
+  ): CoreInterface {
     this.config[key] = value;
     return this;
   }
 
-  public setApiClientService(apiClient: ApiClientInterface): CoreService {
+  public setApiClientService(apiClient: ApiClientInterface): CoreInterface {
     this.service.ApiClientService = apiClient;
     return this;
   }