+++ src/__tests__/Defaults.test.ts
@@ -1,8 +1,12 @@
-import {DrupalSettings, DrupalRole, DrupalError} from '../';
+import {DrupalSettings, DrupalRole, DrupalError} from '..';
 
 test('Test for implemented methods', () => {
   const settings = new DrupalSettings();
   const role = new DrupalRole();
-  expect(() => {settings.get('foo')}).toThrow(DrupalError);
-  expect(() => {role.hasRole('foo')}).toThrow(DrupalError);
+  expect(() => {
+    settings.get('foo');
+  }).toThrow(DrupalError);
+  expect(() => {
+    role.hasRole('foo');
+  }).toThrow(DrupalError);
 });